Developer

+++ TimeWorX/app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -30,12 +30,13 @@ class AuthenticatedSessionController extends Controller
     public function store(LoginRequest $request): RedirectResponse
     {
         $request->authenticate();
-       
         $token = Auth::user()->createToken('timeworx')->plainTextToken;
 
         $request->session()->put('sanctum_token', $token);
 
         $request->session()->regenerate();
+
+        Auth::guard('web')->login(Auth::user());
 
         return redirect()->intended(route('dashboard', absolute: false));
     }
@@ -45,16 +46,19 @@ class AuthenticatedSessionController extends Controller
      */
     public function destroy(Request $request): RedirectResponse
     {
-        
-        Auth::user()->tokens()->delete();
-
         $request->session()->forget('sanctum_token');
-
-        Auth::guard('web')->logout();
 
         $request->session()->invalidate();
 
         $request->session()->regenerateToken();
+        if(Auth::user()->tokens())
+        {
+            Auth::user()->tokens()->delete();
+        }
+        
+        Auth::guard('web')->logout();
+        
+       
 
         return redirect('/');
     }
